0.5.0
hopefully final update. the Movetext trait is cleaned up. well, maybe the trait itself isnt, but the usage is. instead of having an Output type, i switched it up and now theres an agent type. so now Variation implements Movetext, not the old, second "VariationMovetext". that was confusing. now its called VariationAgent but you dont ever need to know about it.